Depressurizer.Core: Add missing Steam licenses game inclusion
This was a huge missing layer that was previous based on the localcache.vdf file.
However Steam has now migrated this part on a separate file named licensecache.

Thanks to https://github.com/wynick27/steam-missing-covers-downloader and to https://github.com/SteamDatabase/Protobufs for providing the logics and data needed to be able to read this file.

Related #10